#632: added tests for IdeToolCommandlet-class
Enhanced the intellij test-ressources with workspace.update.editorconfig as well as setup and update folder for the intellij-tool.
Added a Test that checks if the editorconfig file was copied from the settings repo to the workspace after an ide-tool was run.
Added merger for .editorconfig extension which uses the textMerger.